chore: Add inline_svg gem to Gemfile
This commit adds the `inline_svg` gem to the Gemfile. The gem is added with the version constraint `~> 1.9`. This gem allows for inline SVG usage in the application, which can improve performance and flexibility when working with SVG images. By including this gem in the project's dependencies, the application can take advantage of its features.